Update based_func_alicat.py
We use a wrapper to bypass the asynchronous programming imposed by alicat. 

Note : For larger experiments it might decrease the performance of experiments (theoretically..) needs to be tested. 

Note 2 : Now that we found a way to bypass the asynchronous, we will update the function accordingly, an update is coming soon.